Split data4bert module into multiple processors
